Support for IntQuestions in GUI, switched to representing integer values with longs instead of ints for compatibility with JFormattedTextField

+++ olavTrauschke/QL/src/nl/uva/sea/ql/ast/expr/Int.java
@@ -10,7 +10,7 @@ import nl.uva.sea.ql.generalPurposeVisitors.Visitor;
  * Representation of (literals of) the type int in an AST.
  * 
  * @author Olav Trauschke
- * @version 19-mar-2016
+ * @version 30-mar-2016
  */
 public class Int extends NumericExpr {
     
@@ -19,15 +19,15 @@ public class Int extends NumericExpr {
      */
     public static final int HASH_ORIGIN = 259;
     
-    private final Integer value;
+    private final Long value;
     
     /**
      * Constructor for objects of class <code>Int</code>.
      * 
-     * @param theValue an <code>Integer</code> representing the value of the
+     * @param theValue a <code>Long</code> representing the value of the
      *                  constructed <code>Int</code>
      */
-    public Int(Integer theValue) {
+    public Int(Long theValue) {
         assert theValue != null;
         value = theValue;
     }
